feat: split main js files to modules
