+++ Online-store/src/pages/main/index.html
@@ -189,7 +189,7 @@
     </div>
   </footer>
 
-  <script src="js/script.js"></script>
+  <script src="js/script.js" type="module"></script>
 </body>
 
 </html>
