--- Online-store/src/pages/main/js/modules/module.js
@@ -1,1 +1,0 @@
-console.log('Hey from main/modules/module.js');
