+++ Online-store/src/pages/main/js/modules/update-info.js
@@ -1,0 +1,11 @@
+// same module in each page
+function updateCartSummary(totalProductsSelector, productsLabel, totalPriceSelector, priceLabel) {
+  const totalProducts = document.querySelector(totalProductsSelector);
+  const totalPrice = document.querySelector(totalPriceSelector);
+  const cartArray = JSON.parse(localStorage.getItem('RS-online-cart'));
+
+  totalProducts.textContent = productsLabel + cartArray.reduce((accum, product) => accum + product.count, 0);
+  totalPrice.textContent = priceLabel + cartArray.reduce((accum, product) => accum + product.count * product.price, 0);
+}
+
+export default updateCartSummary;
